services/requirements.txt
MySQL-python
mysql-connector
Flask == 0.10.1
Flask_Inputs == 0.2.0
Flask_RESTful == 0.3.4
jsonschema == 2.5.1
kaptan == 0.5.8
sqlobject == 3.1.0
